Questions delete
die() has been avoided.
stored procedure has been used

+++ web/public/api/questions/delete.php
@@ -8,16 +8,23 @@
 
 		$api_key = isset($_GET['key']) ? $_GET['key'] : die(json_encode(array('message' => 'No API key provided.')));
 
-		$database = new Database();
+		$expected = ['questionID'];
+		$missing = [];
+
+		$database = new Database(false);
 		$db = $database->connect($api_key);
 
 		$questions = new Question($db);
 		
 		$input = json_decode(file_get_contents('php://input'), true);
 
-		$questions->questionID = !empty($input['questionID']) ? $input['questionID'] : die();
+		$questions->questionID = !empty($input['questionID']) ? $input['questionID'] : array_push($missing, 'questionID');
 
-		$questions->delete();
+		if(empty($missing)) {
+			$questions->delete();
+		} else {
+			die(json_encode(array('expected' => $expected, 'missing' => $missing), JSON_PRETTY_PRINT));
+		}
 	} else {
 		echo json_encode(array('message' => 'Wrong HTTP request method. Use DELETE instead.'));
 	}
